Move ESSD logic to a separate module
